\name{print.monet}
\title{Print a monet (monitoring network) object}
\author{C. Gebhardt}
\usage{\method{print}{monet}(x,...)
}
\alias{print.monet}
\arguments{
  \item{x}{object of class \code{"monet"}}
  \item{...}{additional paramters for \code{print}}
}
\description{prints a summary of \code{"x"}
}
\value{None
}
\seealso{
  \code{\link{monet}},
}
\keyword{spatial}
\keyword{ design }